+++ changelog.md
@@ -1,3 +1,7 @@
+# v1.1.1
+    - Allowed use of partially searching usernames. (Like "fi" instead of "Fire")
+    - Fixed Pointercrate issue on levels with spaces (Mac Only)
+    - Fixed issue with friend list searching not working.
 # v1.1.0
     - More Leaderboards added (Diamonds, User Coins, Demons)
     - Added Moderator List